+++ src/discord/mod.rs
@@ -1,6 +1,7 @@
-pub(crate) mod bot_twilight;
+pub(crate) mod bot;
 mod webhooks;
 pub(crate) use webhooks::Webhooks;
+pub(crate) mod discord_state;
 
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
